moving App.js to components, fixing imports
