+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { createStore } from "redux";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./components/App";
 import * as serviceWorker from "./serviceWorker";
 import rootReducer from "./reducers/books";
 import { Provider } from "react-redux";
